+++ src/main/java/blackjack/dto/PlayerProfitResult.java
@@ -1,10 +1,11 @@
 package blackjack.dto;
 
 import blackjack.domain.Player;
+import java.math.BigDecimal;
 
-public record PlayerProfitResult(String playerName, double profit) {
+public record PlayerProfitResult(String playerName, BigDecimal profit) {
 
-    public static PlayerProfitResult from(Player player, Double profit) {
+    public static PlayerProfitResult from(Player player, BigDecimal profit) {
         return new PlayerProfitResult(player.getName(), profit);
     }
 }
